% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterimg.R
\name{renderScatterimg}
\alias{renderScatterimg}
\title{Widget render function for use in Shiny}
\usage{
renderScatterimg(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

