% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterimg.R
\name{scatterimg}
\alias{scatterimg}
\title{<Add Title>}
\usage{
scatterimg(d, opts = NULL, tooltipOpts = NULL, theme = "basic",
  customCSS = NULL, noOverlap = FALSE, noOverlapSpread = NULL,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

