% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterimg.R
\name{scatterimg}
\alias{scatterimg}
\title{<Add Title>}
\usage{
scatterimg(d, fixedCols = NULL, rowLabelCol = NULL, fixedRows = NULL,
  header = NULL, theme = "basic", customCSS = NULL, width = NULL,
  height = NULL)
}
\description{
<Add Description>
}

